doing more script stuff
