+++ otherlib/CMakeLists.txt
@@ -44,6 +44,8 @@ target_sources(otherlib
     src/scripting/binding_utils/bind_glm_matrices.cpp
     src/scripting/binding_utils/bind_math_types_lua.cpp
     src/scripting/binding_utils/bind_rendering_types_lua.cpp
+    src/scripting/dotnet_bindings/environment_api_bindings.cpp
+    src/scripting/dotnet_bindings/ui_bindings.cpp
 
     src/scripting/actions/action.cpp
     src/scripting/actions/callback.cpp
