+++ otherlib/src/other.cpp
@@ -119,8 +119,11 @@ namespace other {
       bind_primary_scripting_environment(config);
       bind_environment_scripts();
     }
+    {
+      PROFILE_SECTION("other::entry--initialize-input-system");
+      subsystem<input_system>::get()->initialize();
+    }
 
-    subsystem<input_system>::get()->initialize();
     /// \note maybe not doing this here anymore
     /// \todo handle other-driver registration here, this includes loading everything not pulled from environment config file
     ///        and registering/initializing all user-facing APIs (this includes things like registering user-facing log, registering user events, etc)
